third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/source-list-parsing-deprecated-expected.txt

CONSOLE ERROR: The 'allow' directive has been replaced with 'default-src'. Please use that directive instead, as 'allow' has no effect.
CONSOLE ERROR: The 'options' directive has been replaced with the 'unsafe-inline' and 'unsafe-eval' source expressions for the 'script-src' and 'style-src' directives. Please use those directives instead, as 'options' has no effect.
CONSOLE ERROR: The 'policy-uri' directive has been removed from the specification. Please specify a complete policy via the Content-Security-Policy header.
Test that directives that predate the CSP 1.0 standard generate proper warnings to set developer expectations correctly.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
PASS

--------
Frame: '<!--framePath //<!--frame2-->-->'
--------
PASS
